% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{uu_colours}
\alias{uu_colours}
\title{UU palette}
\usage{
uu_colours
}
\description{
Utrecht University colour palette, using colours from the house style.
See: https://www.uu.nl/en/organisation/corporate-identity/brand-policy/colour
}
