% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{uu_colours}
\alias{uu_colours}
\title{UU palette}
\format{
An object of class \code{list} of length 2.
}
\usage{
uu_colours
}
\description{
Utrecht University colour palette
}
\keyword{datasets}
